Add a solution for problem 4.
This works within a second or two but I'm convinced this can be
done faster by enumerating palindromes in descending order,
and testing whether each palindrome is a product of three-digit
numbers by factoring it and trying to find a three-digit
product of primes.

I'll check this in as it is for now though.