Move the bulk of the solution for Problem 8 into the MPC module.
'MPC' stands for Maximum Product of Consecutive values. The 'mpc'
function, which used to be called 'mcp', finds the maximum product
of 'k' consecutive values in an arbitrary list of numbers, if the
list is at least 'k' elements long.

I've created a seperate module because this code is useful for
solving Problem 11.